Add the rest of the example site to the project
The Hugo Book instructions only have you pulling the content from the
example site into your new one.  The result is similar to, but not
quite the same as, the theme's example site.

This commit adds all the other bits of the example site to our initial
site, so that it looks like the Real Thing and demo's more of the
functionality.