Add Notes section , w/ a page for workflow diagram
This commit adds a "Notes" section to the left side menu and tucks a
link to a page for the Git workflow diagram underneath ht.

The Git workflow diagram page demonstrates *several* methods for
setting up the link, "someone" should choose one and the others should
be deleted.

Also adds images for the the workflow (small and full-sized) and the
Omnigraffle source file, for future fine tuning.